Refactored OrderCreateCommandHandler code as Transaction annotation would not have worked like that
